+++ CHANGELOG.md
@@ -3,6 +3,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.28] - 2021-04-27
+
+### Added
+- Reset records when user parameters are changed.
+- Minimum and maximum page number in the input.
+
 ## [0.1.27] - 2021-04-22
 
 ### Added